Extract review phase timing, cache events, and budget evidence slices.
Move visible budget projection, prompt cache telemetry, phase timing helpers, and explicit mention lifecycle/publish-failure wiring out of the handlers with unit tests.
